wip(phase06): step 1~5 — input replay reconcile 알고리즘 박힘
- step 1: PDL int→uint 정합(clientTick/lastAckedClientTick) + PacketGenerator uint 지원 + Manager 잠복 버그 (A)안 우회
- step 2: Phase 04 자리잡이로 자연 닫힘 — GameMap.cs (int) 캐스트 제거만
- step 3: InputHistory.cs (sealed, UnityEngine X, 좁은 API 5개) + Unity Test Framework 인프라 첫 박제 (Dawnholder.Client.asmdef + Tests.EditMode.asmdef) + NUnit 12 테스트 + 학습 노트 (int vs uint)
- step 4: 송신 throttle 50ms (Constants.TickDuration, 서버 20 TPS와 1:1 align) + InputHistory wire (PlayerPredictor.NotifySent)
- step 5: OnSnapshot 시그니처 확장 (uint ackedClientTick) + replay reconcile 알고리즘 + smoke test 3/3 통과

step 6 (수동 Play 모드 검증) 진행 대기. .NET 빌드 0 error + 62 test 통과 + Unity 컴파일 0 error.

+++ 02_Server/GameServer/Maps/GameMap.cs
@@ -76,7 +76,7 @@ public class GameMap
                     x = p.Position.X,
                     y = p.Position.Y,
                     serverTick = (int)tickNumber,
-                    lastAckedClientTick = (int)p.LastClientTick
+                    lastAckedClientTick = p.LastClientTick
                 };
                 // 본 Phase는 본인 1명만 — unicast. 다인은 M3+에서 broadcast로 확장.
                 p.Owner.Send(pkt.Write());
